requirements/base.txt
django==1.5.1
psycopg2==2.5
djangorestframework==2.3.9
django-cors-headers==0.12
South==0.8.1
django-compressor==1.3
django-require==1.0.5
Scrapy==0.24.4
tweepy==2.3.0
jusText==2.1.1
